requirements.txt
fastapi==0.112.0
uvicorn==0.30.1
pydantic==2.8.2
pandas==2.2.2
python-dotenv==1.0.1
pyyaml==6.0.2
pytest==8.3.2
